Commands: handle interrupt handler on Windows
Windows does not use nor support the concept of signals.  This adds a
Windows specific path for the interrupt handler to intercept `^C`
requests.  Although this is a bit looser than it should be, this is
equivalent to the previous state.  Unfortunately, due to the need for a
thin thunk, we cannot capture `self` and must use static storage for the
process set and build system ref.